requirements.txt
zope
pytesseract
opencv-python==4.1.0.25
numpy==1.16.4
keyboard==0.13.5
Pillow>=6.2.2
Twisted>=20.3.0
zope.interface==4.4.2
imutils==0.5.3
